implemented car import in db functionality

+++ 8-Spring Data/JSON Processing/CarDealer/src/main/java/bg/softuni/cardealer/data/repositories/PartRepository.java
@@ -3,5 +3,5 @@ package bg.softuni.cardealer.data.repositories;
 import bg.softuni.cardealer.data.entities.Part;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PartRepository extends JpaRepository<Part, Integer> {
+public interface PartRepository extends JpaRepository<Part, Long> {
 }
