+++ 8-Spring Data/JSON Processing/CarDealer/src/main/java/bg/softuni/cardealer/data/repositories/SupplierRepository.java
@@ -3,5 +3,5 @@ package bg.softuni.cardealer.data.repositories;
 import bg.softuni.cardealer.data.entities.Supplier;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SupplierRepository extends JpaRepository<Supplier, Integer> {
+public interface SupplierRepository extends JpaRepository<Supplier, Long> {
 }
